requirements.txt
fastapi
uvicorn
python-multipart
tensorflow==2.12.0
keras==2.12.0
numpy
pillow
requests
python-dotenv
google-generativeai
